requirements.txt
asgiref==3.5.1
Django==4.0.4
sqlparse==0.4.2
tzdata==2022.1
pytest-django
Pillow~=9.1.0
psycopg2~=2.9.3
pip~=21.0.1
attrs~=21.4.0
py~=1.11.0
colorama~=0.4.4
numpy~=1.20.1
setuptools~=49.2.1
djangorestframework~=3.13.1